Create midreview_mack_micah
Mack's first submission of the mid-review code. This was tested at home with the 8 GLOW LED strip. There are two basic buttons, one being the master on/off button, and the other being what detects the signal. This will be replaced with the sinulator when the time comes. The two pots, one is for hue and the other is for intensity. 

Before any major testing with more LEDs, there are a few items that need to be tweaked in the code, beginning with the NUM_LEDS. 

There is room, and hopefully it is clear enough to Micah what I did for him to add his sound code. I have added his input to the define function.